Move all validations to pydantic schemas
